+++ BitFaster.Caching.Benchmarks/BitFaster.Caching.Benchmarks.csproj
@@ -9,6 +9,8 @@
     <IsWindows Condition="'$([System.Runtime.InteropServices.RuntimeInformation]::IsOSPlatform($([System.Runtime.InteropServices.OSPlatform]::Windows)))' == 'true'">true</IsWindows> 
     <IsLinux Condition="'$([System.Runtime.InteropServices.RuntimeInformation]::IsOSPlatform($([System.Runtime.InteropServices.OSPlatform]::Linux)))' == 'true'">true</IsLinux>
     <IsMacOS Condition="'$([System.Runtime.InteropServices.RuntimeInformation]::IsOSPlatform($([System.Runtime.InteropServices.OSPlatform]::OSX)))' == 'true'">true</IsMacOS>
+    <IsArm64 Condition="$([System.Runtime.InteropServices.RuntimeInformation]::OSArchitecture) == Arm64">true</IsArm64>
+    <IsX64 Condition="$([System.Runtime.InteropServices.RuntimeInformation]::OSArchitecture) == X64">true</IsX64>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
@@ -40,6 +42,12 @@
     </PropertyGroup>
     <PropertyGroup Condition="'$(IsMacOS)'=='true'">
         <DefineConstants>MacOS</DefineConstants>
+    </PropertyGroup>
+    <PropertyGroup Condition="'$(IsArm64)'=='true'">
+        <DefineConstants>Arm64</DefineConstants>
+    </PropertyGroup>
+    <PropertyGroup Condition="'$(IsX64)'=='true'">
+        <DefineConstants>X64</DefineConstants>
     </PropertyGroup>
 
 </Project>
